backend/requirements.txt
# OurColumbus Backend Dependencies

# Web scraping
playwright>=1.40.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
supabase>=2.0.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
